+++ jupyterhub/spawners.py
@@ -21,6 +21,7 @@ import os
 from urllib.parse import urlsplit, urlunsplit, urlparse
 
 import escapism
+import gitlab
 from kubernetes import client
 from tornado import gen, web
 
@@ -89,8 +90,6 @@ class SpawnerMixin:
     @gen.coroutine
     def start(self, *args, **kwargs):
         """Start the notebook server."""
-        import gitlab
-
         self.log.info("starting with args: {}".format(" ".join(self.get_args())))
         self.log.debug("user options: {}".format(self.user_options))
 
@@ -142,6 +141,7 @@ class SpawnerMixin:
             access_level = gitlab.GUEST_ACCESS
             if len(access_levels) > 0:
                 access_level = max(access_levels)
+            self.gl_access_level = access_level
 
             if access_level >= gitlab.MAINTAINER_ACCESS:
 
@@ -263,6 +263,18 @@ class RenkuKubeSpawner(SpawnerMixin, KubeSpawner):
             }
         }
 
+        # 6. Set up the https proxy for GitLab
+        https_proxy = client.V1Container(
+            name="git-https-proxy",
+            env=[
+                client.V1EnvVar(name="GITLAB_OAUTH_TOKEN", value=oauth_token),
+                client.V1EnvVar(name="REPOSITORY_URL", value=repository),
+                client.V1EnvVar(name="MITM_PROXY_PORT", value="8080"),
+            ],
+            image=options.get("git_https_proxy_image"),
+        )
+        self.extra_containers.append(https_proxy)
+
         # Finalize the pod configuration
 
         # Set the repository path to the working directory
@@ -281,24 +293,23 @@ class RenkuKubeSpawner(SpawnerMixin, KubeSpawner):
             os.environ.get("GITLAB_URL", "http://gitlab.renku.build"),
         )
         git_host = parsed_git_url.netloc
+        safe_username = escapism.escape(self.user.name, escape_char="-").lower()
         self.extra_annotations = {
             RENKU_ANNOTATION_PREFIX + "namespace": options.get("namespace"),
-            RENKU_ANNOTATION_PREFIX + "projectName": options.get("project"),
             RENKU_ANNOTATION_PREFIX
             + "projectId": "{}".format(options.get("project_id")),
             RENKU_ANNOTATION_PREFIX + "branch": options.get("branch"),
-            RENKU_ANNOTATION_PREFIX + "commit-sha": options.get("commit_sha"),
             RENKU_ANNOTATION_PREFIX + "repository": repository_url,
+            RENKU_ANNOTATION_PREFIX + "git-host": git_host,
+            RENKU_ANNOTATION_PREFIX + "username": safe_username,
+            RENKU_ANNOTATION_PREFIX + "commit-sha": options.get("commit_sha"),
+            RENKU_ANNOTATION_PREFIX + "projectName": options.get("project"),
         }
-
-        # add username to labels
-        safe_username = escapism.escape(self.user.name, escape_char="-").lower()
+        # some annotations are repeated as labels so that the k8s api can filter resources
         self.extra_labels = {
             RENKU_ANNOTATION_PREFIX + "username": safe_username,
             RENKU_ANNOTATION_PREFIX + "commit-sha": options.get("commit_sha"),
             RENKU_ANNOTATION_PREFIX + "projectName": options.get("project"),
-            RENKU_ANNOTATION_PREFIX + "git-host": git_host,
-            RENKU_ANNOTATION_PREFIX + "namespace": options.get("namespace"),
         }
 
         self.delete_grace_period = 30