+++ renku_notebooks/wsgi.py
@@ -16,6 +16,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Notebooks service app wrapper."""
-from renku_notebooks import create_app
+from renku_notebooks import create_app, register_swagger
 
 app = create_app()
+app = register_swagger(app)